Add fronze state to actor
Actor.freeze()
The actor will be kept on the screen but won't be animated.
Has the same effect as setting Actor.animate to an empty function,
except the animate function doesn't get lost in the process.
It also reads better.
@return this

Actor.unfreeze()
Removes actor from frozen state and resumes animation.
@return this